backend/requirements.txt
fastapi
uvicorn
ollama
requests
python-dotenv
slowapi
googlemaps